addressbook/ui: moved TextUI.java format methods into Formatter.java
